Nest interpolation back into Tempo
