+++ RhythmTests/EasingEvaluateTests.swift
@@ -15,7 +15,7 @@ class EasingEvaluateTests: XCTestCase {
     func testLinear() {
 
         let x: Double = 0.5
-        let ease = Interpolation.Easing.linear
+        let ease = Tempo.Interpolation.Easing.linear
         let expected = x
         let result = ease.evaluate(at: x)
         XCTAssertEqual(result, expected)
@@ -26,7 +26,7 @@ class EasingEvaluateTests: XCTestCase {
     func testPowerInOne() {
 
         let x: Double = 0.5
-        let ease = Interpolation.Easing.powerIn(exponent: 1)
+        let ease = Tempo.Interpolation.Easing.powerIn(exponent: 1)
         let expected = 0.5
         let result = ease.evaluate(at: x)
         XCTAssertEqual(result, expected)
@@ -35,7 +35,7 @@ class EasingEvaluateTests: XCTestCase {
     func testPowerInTwo() {
 
         let x: Double = 0.5
-        let ease = Interpolation.Easing.powerIn(exponent: 2)
+        let ease = Tempo.Interpolation.Easing.powerIn(exponent: 2)
         let expected = 0.25
         let result = ease.evaluate(at: x)
         XCTAssertEqual(result, expected)
@@ -44,7 +44,7 @@ class EasingEvaluateTests: XCTestCase {
     func testPowerInHalf() {
 
         let x: Double = 0.5
-        let ease = Interpolation.Easing.powerIn(exponent: 0.5)
+        let ease = Tempo.Interpolation.Easing.powerIn(exponent: 0.5)
         let expected = 0.70710678118 // 1 / sqrt(2)
         let result = ease.evaluate(at: x)
         XCTAssertEqualWithAccuracy(result, expected, accuracy: 1e-6)
@@ -55,7 +55,7 @@ class EasingEvaluateTests: XCTestCase {
     func testPowerInOutOne() {
 
         let x: Double = 0.25
-        let ease = Interpolation.Easing.powerInOut(exponent: 1)
+        let ease = Tempo.Interpolation.Easing.powerInOut(exponent: 1)
         let expected = 0.25
         let result = ease.evaluate(at: x)
         XCTAssertEqual(result, expected)
@@ -63,7 +63,7 @@ class EasingEvaluateTests: XCTestCase {
 
     func testPowerInOutTwo() {
 
-        let ease = Interpolation.Easing.powerInOut(exponent: 2)
+        let ease = Tempo.Interpolation.Easing.powerInOut(exponent: 2)
 
         XCTAssertEqual(ease.evaluate(at: 0), 0)
         XCTAssertEqual(ease.evaluate(at: 0.25), 0.125)
@@ -74,7 +74,7 @@ class EasingEvaluateTests: XCTestCase {
 
     func testPowerInOutThree() {
 
-        let ease = Interpolation.Easing.powerInOut(exponent: 3)
+        let ease = Tempo.Interpolation.Easing.powerInOut(exponent: 3)
         XCTAssertEqual(ease.evaluate(at: 0), 0)
         XCTAssertEqual(ease.evaluate(at: 0.25), 0.0625)
         XCTAssertEqual(ease.evaluate(at: 0.5), 0.5)
@@ -85,7 +85,7 @@ class EasingEvaluateTests: XCTestCase {
     // MARK: - SineInOut
 
     func testSineInOut() {
-        let ease = Interpolation.Easing.sineInOut
+        let ease = Tempo.Interpolation.Easing.sineInOut
         let inputs: [Double] = [0, 0.25, 0.5, 0.75, 1]
         let sqrt2_recip = 1 / sqrt(2)
         let expecteds: [Double] = [
